% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raukrtemplate.R
\name{create_template}
\alias{create_template}
\title{Create an Rmd template}
\usage{
create_template(template = NULL, path = NULL, name = NULL, verbose = TRUE)
}
\arguments{
\item{template}{[Character] denoting name of the template. See details.}

\item{path}{[Character] A directory path with write permissions where the template will be created}

\item{name}{[Character] An optional character denoting name of the template document and it's parent directory. Defaults to same value as "template".}

\item{verbose}{[Logical] indicating if message is printed to console.}
}
\value{
None
}
\description{
Create an Rmd template in a specified directory
}
\details{
Possible options for argument "template" are "lab", "lab-demo", "presentation", "presentation-demo".
}
\examples{
\dontrun{
create_template("lab", "~/Desktop")
create_template("lab", "~/Desktop", "my-topic")
}

}
